float alarm

+++ InvisibleTutorial/Enums/Message.cs
@@ -8,7 +8,7 @@ namespace InvisibleTutorial.Enums
         public string Content { get; set; }
 
         [Description("The duration of the message.")]
-        public ushort Duration { get; set; }
+        public float Duration { get; set; }
 
         [Description("A bool indicating whether to show this message or not.")]
         public bool Show { get; set; }
